requirements.txt
Flask
python-dotenv
pdf2image
Pillow
fitz
frontend
tools
python-pptx
python-dateutil